Added utilities class for LibPlcTag. Renamed dll import class filename to match the class name.
